Fix issues reported by PREfast static analysis tool (dotnet/coreclr#24577)
Commit migrated from https://github.com/dotnet/coreclr/commit/718598334310f6ad331ba2893ee4557beffaa74c

+++ src/coreclr/src/gc/env/gcenv.windows.inl
@@ -10,7 +10,7 @@
 
 #define OS_PAGE_SIZE GCToOSInterface::GetPageSize()
 
-__forceinline uint32_t GCToOSInterface::GetPageSize()
+__forceinline size_t GCToOSInterface::GetPageSize()
 {
     return 0x1000;
 }
